refactore/KD-53 : 기존 aics-api -> aics-admin으로 파일 이동시켰습니다.
