+++ aics-api/src/main/java/kgu/developers/api/certificate/presentation/CertificateControllerImpl.java
@@ -3,14 +3,12 @@ package kgu.developers.api.certificate.presentation;
 import static org.springframework.http.MediaType.MULTIPART_FORM_DATA_VALUE;
 
 import kgu.developers.api.certificate.application.CertificateFacade;
-import kgu.developers.api.certificate.presentation.response.CertificateDetailResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import kgu.developers.api.certificate.presentation.request.CertificateSubmitRequest;
 import kgu.developers.api.certificate.presentation.response.CertificatePersistResponse;
-import kgu.developers.domain.certificate.application.command.CertificateCommandService;
 import lombok.RequiredArgsConstructor;
 
 @RestController
@@ -30,11 +28,4 @@ public class CertificateControllerImpl implements CertificateController {
 			CertificatePersistResponse.of(id)
 		);
 	}
-	@Override
-	@GetMapping("/{id}")
-	public ResponseEntity<CertificateDetailResponse> getCertificate(@PathVariable Long id) {
-		return ResponseEntity.ok(certificateFacade.getById(id));
-	}
-
-
 }
