+++ aics-api/src/main/java/kgu/developers/api/thesis/application/ThesisFacade.java
@@ -1,15 +1,11 @@
 package kgu.developers.api.thesis.application;
 
-import kgu.developers.api.thesis.presentation.response.ThesisDetailResponse;
-import kgu.developers.domain.file.application.query.FileQueryService;
 import kgu.developers.domain.graduationUser.application.command.GraduationUserCommandService;
 import kgu.developers.domain.graduationUser.application.query.GraduationUserQueryService;
 import kgu.developers.domain.graduationUser.domain.GraduationUser;
 import kgu.developers.domain.schedule.application.query.ScheduleQueryService;
 import kgu.developers.domain.schedule.domain.Schedule;
 import kgu.developers.domain.thesis.application.command.ThesisCommandService;
-import kgu.developers.domain.thesis.application.query.ThesisQueryService;
-import kgu.developers.domain.thesis.domain.Thesis;
 import kgu.developers.domain.user.application.query.UserQueryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -24,8 +20,6 @@ public class ThesisFacade {
     private final GraduationUserQueryService graduationUserQueryService;
     private final ScheduleQueryService scheduleQueryService;
     private final GraduationUserCommandService graduationUserCommandService;
-    private final ThesisQueryService thesisQueryService;
-    private final FileQueryService fileQueryService;
 
     public Long submitThesis(MultipartFile file, Long scheduleId) {
         Long thesisId = thesisCommandService.submitThesis(file,scheduleId);
@@ -37,11 +31,5 @@ public class ThesisFacade {
         graduationUserCommandService.updateThesis(graduationUser, thesisId, schedule);
         return thesisId;
     }
-    public ThesisDetailResponse getById(Long id){
-        Thesis thesis = thesisQueryService.getById(id);
-        String physicalPath = thesis.getThesisFileId() != null
-                ? fileQueryService.getFilePhysicalPath(thesis.getThesisFileId())
-                : null;
-        return ThesisDetailResponse.from(thesis, physicalPath);
-    }
+
 }
