+++ aics-common/src/main/java/kgu/developers/common/config/SecurityConfig.java
@@ -86,8 +86,6 @@ public class SecurityConfig {
 		"/api/v1/thesis",
 		"/api/v1/certificate",
 		"/api/v1/schedules",
-		"/api/v1/thesis/**",
-		"/api/v1/certificate/**",
 	};
 
 	CorsConfigurationSource corsConfigurationSource() {
